+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 - Multihop: changes total_commission_amount type in SimulateSwap and SimulateReverseSwap ([#236])
 - Multihop: removes unnecessary unwrap of a value ([#240])
 - Factory, Multihop, Pool, Pool_stable, Phoenix: adds lp_token decimal's as a const instead a user input ([#241])
+- Factory, Multihop, Pool, Pool_stable: adds a new parameter for creating liquidity pool ([#243])
 
 [#200]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/200
 [#234]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/234
@@ -26,12 +27,15 @@ and this project adheres to
 [#236]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/236
 [#240]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/240
 [#241]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/241
+[#243]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/243
 
 ## Added
 
 - Adds a new macro that validates the bps arguments value ([#199])
+- Added test coverage for Decimal ([#244])
 
 [#199]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/199
+[#244]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/244
 
 ## Bug fixes